prevPred/requirements.txt
flask==3.0.3
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
requests==2.31.0
gunicorn
PyMySQL==1.1.1
SQLAlchemy==2.0.30
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
